Petsc configure helper: convert to python
This will hopefully be easier to work with and won't have the
huge bug that the previous script had, of swallowing the opt flags.

This intends to be a lighter wrapper around the PETSc configure script:
it accepts all the normal arguments for that script, as well as its own,
and processes all of them to create the final set of options and a
PETSC_ARCH value.
.

.

.